requirements.txt
annotated-types==0.7.0
anyio==4.4.0
apt-clone==0.2.1
apturl==0.5.2
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
click==8.1.7
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.6.1
duplicity==0.8.12.0
email-validator==2.2.0
entrypoints==0.3
exceptiongroup==1.2.2
fastapi==0.112.0
fastapi-cli==0.0.5
fasteners==0.14.1
future==0.18.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.14.0
httptools==0.6.1
httpx==0.27.0
idna==2.8
jinja2==3.1.4
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
monotonic==1.5
motor==3.5.1
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
PAM==0.4.2
paramiko==2.6.0
passlib==1.7.4
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
pycairo==1.16.2
pycups==1.9.73
pydantic==2.8.2
pydantic-core==2.20.1
pygments==2.18.0
PyGObject==3.36.0
PyICU==2.4.2
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==4.8.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==1.0.1
python-http-client==3.3.7
python-multipart==0.0.9
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rich==13.7.1
SecretStorage==2.3.1
sendgrid==6.11.0
shellingham==1.5.4
simplejson==3.16.0
six==1.14.0
sniffio==1.3.1
starkbank-ecdsa==2.2.0
starlette==0.37.2
systemd-python==234
typer==0.12.3
typing-extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
uvicorn==0.30.5
uvloop==0.19.0
wadllib==1.3.3
watchfiles==0.22.0
websockets==12.0
xkit==0.0.0
